Move hashing functionality to ArchiveModelManager
